refactor(agents-api): move strip_embeddings util to get_doc_search module
